refactoring: HeadBar
