--- view/desktop/src/components/HeadBar/index.ts
@@ -1,1 +1,0 @@
-export * from "./HeadBar";
